use bigint for deserialization issues

+++ api/src/chromia/chromia.dtos.ts
@@ -1,6 +1,6 @@
 interface player_asset_info {
   name: string;
-  amount: BigInteger;
+  amount: bigint;
   [key: string]: any; // Index signature for GTV compatibility
 }
 interface shop_listing {
